[Feat] Display paginated data
