+++ components/shared/Pagination.js
@@ -2,14 +2,14 @@ import ReactPaginate from 'react-paginate';
 
 import styles from 'styles/Pagination.module.css';
 
-const Pagination = ({ totalElements = 0, itemsPerPage = 5 }) => {
+const Pagination = ({ totalElements = 0, pageSize = 5, onPageChange }) => {
   return (
     <div className={styles.pagination}>
       <ReactPaginate
         breakLabel="..."
-        onPageChange={() => {}}
+        onPageChange={({ selected }) => onPageChange(selected + 1, pageSize)}
         pageRangeDisplayed={5}
-        pageCount={Math.ceil(totalElements / itemsPerPage)}
+        pageCount={Math.ceil(totalElements / pageSize)}
         nextLabel="&rarr;"
         previousLabel="&larr;"
         renderOnZeroPageCount={null}
