+++ pages/forum/topics/[slug].js
@@ -1,16 +1,18 @@
-import { useState, useRef } from 'react';
+import { useState } from 'react';
 import { useRouter } from 'next/router';
 
-import PostItem from 'components/forum/PostItem';
-import Replier from 'components/shared/Replier';
-import Pagination from 'components/shared/Pagination';
+// Components
+import Posts from 'components/forum/Posts';
+
+// Hooks
 import {
   useGetPostsByTopic,
   useGetTopicBySlug,
-  useGetUser,
-  useCreatePost
+  useGetUser
 } from 'apollo/hooks';
-import WithApollo from 'hoc/withApollo';
+
+// Misc
+import withApollo from 'hoc/withApollo';
 
 const useInitialData = (pagination) => {
   const router = useRouter();
@@ -21,7 +23,7 @@ const useInitialData = (pagination) => {
   });
   const topic = (dataT && dataT.topicBySlug) || null;
 
-  const { data: dataP } = useGetPostsByTopic({
+  const { data: dataP, fetchMore } = useGetPostsByTopic({
     variables: { slug, ...pagination }
   });
   const postsData = (dataP && dataP.postsByTopic) || { content: [] };
@@ -29,12 +31,22 @@ const useInitialData = (pagination) => {
   const { data: dataU } = useGetUser();
   const user = (dataU && dataU.user) || null;
 
-  return { topic, ...postsData, user };
+  return { topic, ...postsData, user, fetchMore };
 };
 
 const PostsPage = () => {
-  const [pagination, setPagination] = useState({ offset: 0, limit: 5 });
-  const { topic, content, ...rest } = useInitialData(pagination);
+  const [pagination, setPagination] = useState({ pageNum: 1, pageSize: 5 });
+  const { topic, content, fetchMore, ...rest } = useInitialData(pagination);
+
+  const handlePageChange = (pageNum, pageSize) => {
+    setPagination({ pageNum, pageSize });
+    fetchMore({
+      variables: {
+        pageNum,
+        pageSize
+      }
+    });
+  };
 
   return (
     topic && (
@@ -46,95 +58,16 @@ const PostsPage = () => {
             </div>
           </div>
         </section>
-        <Posts topic={topic} posts={content} {...pagination} {...rest} />
+        <Posts
+          topic={topic}
+          posts={content}
+          {...pagination}
+          {...rest}
+          onPageChange={handlePageChange}
+        />
       </>
     )
   );
 };
 
-const Posts = ({ topic, posts, user, totalElements, offset, limit }) => {
-  const pageEnd = useRef(null);
-  const [isReplierOpen, setReplierOpen] = useState(false);
-  const [replyTo, setReplyTo] = useState(null);
-
-  const [createPost, { error }] = useCreatePost();
-
-  const handleCreate = () => {
-    setReplyTo(null);
-    setReplierOpen(true);
-  };
-
-  const handleReply = (reply) => {
-    setReplyTo(reply);
-    setReplierOpen(true);
-  };
-
-  const scrollToBottom = () =>
-    pageEnd.current.scrollIntoView({ behavior: 'smooth' });
-
-  const handleSubmit = async (data, resetReplier) => {
-    if (!data.content) return null;
-
-    const reply = {
-      content: data.content,
-      topic: topic._id
-    };
-
-    if (replyTo) {
-      reply.parent = replyTo._id;
-    }
-
-    await createPost({ variables: { input: reply } });
-    resetReplier();
-    setReplierOpen(false);
-    scrollToBottom();
-  };
-
-  return (
-    <section className="pb-5">
-      <div className="fj-post-list">
-        <PostItem post={topic} className="topic-post-lead" />
-        {posts.length > 0 &&
-          posts.map((post) => (
-            <div className="row" key={post.slug}>
-              <div className="col-md-9">
-                <PostItem
-                  post={post}
-                  onReply={handleReply}
-                  canCreate={!!user}
-                />
-              </div>
-            </div>
-          ))}
-      </div>
-      <div className="row mt-2 mx-0">
-        <div className="col-md-9">
-          <div>
-            {user ? (
-              <div className="pt-2 pb-2">
-                <button
-                  className="btn btn-lg btn-outline-primary"
-                  type="button"
-                  onClick={handleCreate}
-                >
-                  Create New Post
-                </button>
-              </div>
-            ) : null}
-          </div>
-          <Pagination totalElements={totalElements} itemsPerPage={5} />
-        </div>
-      </div>
-      <div ref={pageEnd} />
-      <Replier
-        hasTitle={false}
-        isOpen={isReplierOpen}
-        replyTo={(replyTo && replyTo.user.username) || topic.title}
-        onClose={() => setReplierOpen(false)}
-        onSubmit={handleSubmit}
-      />
-    </section>
-  );
-};
-
-export default WithApollo(PostsPage);
+export default withApollo(PostsPage);
